upload sketches(PlanQk-85): implement: getSketchImage, getSketch and getSketches
