+++ org.planqk.atlas.core/src/main/java/org/planqk/atlas/core/repository/ImageRepository.java
@@ -4,6 +4,8 @@ import java.util.UUID;
 
 import org.planqk.atlas.core.model.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 /**
@@ -11,5 +13,8 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
  */
 @RepositoryRestResource(exported = false)
 public interface ImageRepository extends JpaRepository<Image, UUID> {
+
+    @Query("SELECT i FROM Image i where i.sketch.id = :sketchId")
+    Image findImageBySketchId(@Param("sketchId") UUID sketchId);
 
 }
