+++ org.planqk.atlas.core/src/main/java/org/planqk/atlas/core/repository/SketchRepository.java
@@ -1,9 +1,12 @@
 package org.planqk.atlas.core.repository;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.planqk.atlas.core.model.Sketch;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 /**
@@ -11,5 +14,8 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
  */
 @RepositoryRestResource(exported = false)
 public interface SketchRepository extends JpaRepository<Sketch, UUID> {
+
+    @Query("SELECT s FROM Sketch s where s.algorithm.id = :algorithmId")
+    List<Sketch> findSketchesByAlgorithm_Id(@Param("algorithmId") UUID algorithmId);
 
 }
