+++ org.planqk.atlas.core/src/main/java/org/planqk/atlas/core/services/SketchService.java
@@ -1,5 +1,6 @@
 package org.planqk.atlas.core.services;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.planqk.atlas.core.model.Sketch;
@@ -7,9 +8,15 @@ import org.springframework.web.multipart.MultipartFile;
 
 public interface SketchService {
 
+    public List<Sketch> findByAlgorithm(UUID algorithmId);
+
     public Sketch addSketchToAlgorithm(UUID algorithmId, MultipartFile file, String description, String baseURL);
 
     public void delete(UUID sketchId);
 
     public Sketch findById(UUID id);
+
+    Sketch getSketchByAlgorithmAndSketch(final UUID algoId, final UUID sketchId);
+
+    byte[] getImageByAlgorithmAndSketch(final UUID algoId, final UUID sketchId);
 }
